NeuroStream-Reflex x DeepSeek-V2-Lite-Chat 嫁接代码包（v3.21，替换旧内容）
- 发布运行必需代码（config/core/loop/interaction/learn/improve + 入口脚本）
- 排除：docs、_dsv2_ref、训练文件（improve/self_play.py）、缓存
- README: 权重/tokenizer 指向 ModelScope Marshauv/Neurostream_reflex_deepseek_chat
- 许可: MIT（含 DeepSeek-V2-Lite-Chat 上游 MIT 声明）

+++ learn/critic.py
@@ -41,7 +41,9 @@ class ReflexCritic(nn.Module):
     def forward(self, state_vector):
         if state_vector.dim() == 1:
             state_vector = state_vector.unsqueeze(0)
-        x = self.input_ln(state_vector.float())
+        # 不强制 float：RMSNorm 内部按输入 dtype 计算（bf16 嫁接模式下
+        # 强制 fp32 会导致后续 Linear(bf16) 与 fp32 输入 dtype 不匹配）
+        x = self.input_ln(state_vector)
         return self.net(x)
 
     def get_normalized_v(self, state_vector):
